Rename batch_summaries
