+++ src/remember/summaries.py
@@ -44,7 +44,7 @@ class Summaries(Remembering):
         self.__memory_prompt: str = config.memory_prompt
         self.__resummarize_prompt:str = config.resummarize_prompt
 
-    @utils.time_it
+     @utils.time_it
     def get_prompt_text(self, npcs_in_conversation: Characters, world_id: str) -> str:
         """Load the conversation summaries for all NPCs in the conversation and returns them as one string
 
@@ -122,7 +122,7 @@ class Summaries(Remembering):
             if paragraphs:
                 return "\n".join(paragraphs)
         return ""
-    
+
     @utils.time_it
     def may_add_missing_join_leave_messages(self, messages: message_thread) -> bool:
         """ Adds missing join and leave messages to the beginning / end of the message thread."""
@@ -149,34 +149,31 @@ class Summaries(Remembering):
         if len(characters_joined) < len(characters_found):
             for name, character in characters_found.items():
                 if name not in characters_joined:
-                    messages.insert_after_system_messages(join_message(character, self.__config))
+                    messages.insert_after_system_messages(join_message(character))
                     hadMissingMessages = True
         if len(characters_left) < len(characters_found):
             for name, character in characters_found.items():
                 if name not in characters_left:
-                    messages.add_message(leave_message(character, self.__config))
+                    messages.add_message(leave_message(character))
                     hadMissingMessages = True
         return hadMissingMessages
-
+    
     @utils.time_it
     def save_conversation_state(self, messages: message_thread, npcs_in_conversation: Characters, world_id: str, is_reload=False):
         summary = ''
-        # If we truncated the conversation due to running out of context, join and leave messages may be missing
         self.may_add_missing_join_leave_messages(messages)
         
         characters = self.get_character_lookup_dict(messages)
-        npc_message_threads: Dict[str, CharacterSummaryParameters] = self.get_threads_for_summarization(messages, characters)
+        npc_message_threads = self.get_threads_for_summarization(messages)
         npcs_with_shared_threads = self.group_shared_threads(npc_message_threads)
-        
-        for npc_names in npcs_with_shared_threads:
-           summary = self.__create_new_conversation_summary(npc_message_threads[npc_names[0]], world_id)
-           for npc_name in npc_names:
-               self.__append_new_conversation_summary(summary, characters[npc_name], world_id)
-        
-        
-                
+        for thread, npc_names in npcs_with_shared_threads.items():
+            summary = self.__create_new_conversation_summary(thread, npc_names[0])
+            for npc_names in npc_names:
+                self.__append_new_conversation_summary(summary, characters[npc_names], world_id)
 
-    def get_character_lookup_dict(self, all_messages: message_thread) -> Dict[str, Character]:
+
+        
+    def get_character_lookup_dict(self, all_messages: message_thread) -> dict[str, Character]:
         """Returns a dictionary of character names to Character objects."""
         characters = {}
         for message in all_messages.get_messages_of_type((join_message)):
@@ -184,19 +181,15 @@ class Summaries(Remembering):
                 continue
             characters[message.character.name] = message.character
         return characters
-
+ 
     @utils.time_it
-    def get_threads_for_summarization(self, all_messages: message_thread, characters:Dict[str, Character]) -> Dict[str, CharacterSummaryParameters]:
-        """
-        Returns a dictionary mapping an NPC's name to a CharacterSummaryParameters object,
-        which encapsulates the npc's message_thread and the list of Characters they've seen.
-        """
-        npcs_in_conversation: Dict[str, bool] = {}
+    def get_threads_for_summarization(self, all_messages: message_thread) -> dict[str, message_thread]:
+        """Returns dict[npc name, messagesForThatNpc]"""
+        npcs_in_conversation: dict[str, bool] = {}
         def set_in_conversation(npc: Character, in_conversation: bool):
             npcs_in_conversation[npc.name] = in_conversation
         
-        npc_messageThreads: Dict[str, message_thread] = {}
-        npc_has_seen_npcs: Dict[str, Dict[str, Character]] = {}
+        npc_messageThreads: dict[str, message_thread] = {}
 
         for message in all_messages.get_persistent_messages():
             # Mark npc as present when they join
@@ -205,25 +198,16 @@ class Summaries(Remembering):
 
             # Add the message for each npc that was in the conversation to hear this message
             for npc_name, in_conversation in npcs_in_conversation.items():
-                # For each npc we extract a list of all the other npcs that have been in the conversation with them at the same time
-                for npc_name2, in_conversation2 in npcs_in_conversation.items():
-                    if in_conversation and in_conversation2:
-                        if npc_has_seen_npcs.get(npc_name) is None:
-                            npc_has_seen_npcs[npc_name] = {}
-                        npc_has_seen_npcs[npc_name][npc_name2] = characters[npc_name2]  # Assuming message.character represents the Character.
-                
-                # We also store the message for the npc if they are in the conversation for it
                 if in_conversation:
                     if npc_name not in npc_messageThreads:
-                        npc_messageThreads[npc_name] = message_thread(self.__config, None)
+                        npc_messageThreads[npc_name] = message_thread(None)
                     thread: message_thread = npc_messageThreads[npc_name]
                     
                     # Mark passage of time, in case a character left and rejoined the conversation
-                    if len(thread) > 0:
+                    if thread.__len__() > 0:
                         npcs_previous_message = thread.get_last_message()
                         if isinstance(npcs_previous_message, leave_message) and npcs_previous_message.character.name == npc_name:
-                            narration_start, narration_end = self.__config.get_narration_indicators()
-                            thread.add_message(user_message(self.__config, narration_start + "some time later*" + narration_end))
+                            thread.add_message(assistant_message("* some time later *"))
                     
                     thread.add_message(message)
             
@@ -231,18 +215,9 @@ class Summaries(Remembering):
             if isinstance(message, leave_message) and not message.character.is_player_character:
                 set_in_conversation(message.character, False)
 
-        # Prepare the result
-        result: Dict[str, CharacterSummaryParameters] = {}
-        for npc_name, seen_npcs_dict in npc_has_seen_npcs.items():
-            if npc_name not in npc_messageThreads:
-                continue
-            seen_npcs = [seen_npcs_dict[key] for key in seen_npcs_dict]
-            thread = npc_messageThreads[npc_name]
-            result[npc_name] = CharacterSummaryParameters(thread, seen_npcs)
-            
-        return result
+        return npc_messageThreads
 
-    def group_shared_threads(self, npc_threads: Dict[str, CharacterSummaryParameters]) -> list[list[str]]:
+    def group_shared_threads(self, npc_threads: Dict[str, message_thread]) -> Dict[message_thread, List[str]]:
         """
         Groups NPC message threads if they have exactly the same messages.
 
@@ -254,16 +229,19 @@ class Summaries(Remembering):
         """
         # Group NPCs by the exact tuple of message texts from their thread.
         thread_groups = defaultdict(list)
-        for npc_name, summary_params in npc_threads.items():
-            messages_tuple = tuple(message.text for message in summary_params.messages.get_talk_only())
+        for npc_name, thread in npc_threads.items():
+            messages_tuple = tuple(message.text for message in thread.get_talk_only())
             thread_groups[messages_tuple].append(npc_name)
 
         # Build the result: for each group, select a representative thread.
-        result: list[list[str]] = []
+        result: Dict[message_thread, List[str]] = {}
         for messages, npc_list in thread_groups.items():
-            result.append(npc_list)
+            representative_npc_name = npc_list[0]
+            representative_thread = npc_threads[representative_npc_name]
+            result[representative_thread] = npc_list
 
         return result
+
 
                         
 
@@ -350,8 +328,8 @@ class Summaries(Remembering):
                 )
         while True:
             try:
-                if len(npcInfo.messages) >= 5:
-                    return self.summarize_conversation(npcInfo.messages.transform_to_dict_representation(npcInfo.messages.get_talk_only()), prompt)
+                if len(messages) >= 1:
+                    return self.summarize_conversation(messages.transform_to_dict_representation(messages.get_talk_only()), prompt, npc_name)
                 else:
                     logging.info(f"Conversation summary not saved. Not enough dialogue spoken.")
                 break
@@ -431,14 +409,12 @@ class Summaries(Remembering):
     @utils.time_it
     def summarize_conversation(self, text_to_summarize: str, prompt: str) -> str:
         summary = ''
-        if len(text_to_summarize) > 5:
-            messages = message_thread(self.__config, prompt)
-            messages.add_message(UserMessage(self.__config, text_to_summarize))
-
+        if len(text_to_summarize) > 0:
+            messages = message_thread(prompt)
+            messages.add_message(UserMessage(text_to_summarize))
             # Log the summary prompt being sent
             logging.log(23, f'Summary prompt sent to LLM: {prompt.strip()}')
-
-            summary = self.__summary_client.request_call(messages)
+            summary = self.__client.request_call(messages)
             if not summary:
                 logging.info(f"Summarizing conversation failed.")
                 return ""
